photos refactor
