--- routes/api/fileUploadRoutes.js
@@ -1,108 +1,0 @@
-require("dotenv").config();
-const express = require("express");
-const router = express.Router();
-const DOCUMENT = require("../../models/Document");
-const multer = require("multer");
-let AWS = require("aws-sdk");
-
-let storage = multer.memoryStorage();
-let upload = multer({ storage: storage });
-
-// INDEX ROUTE
-router.route("/").get((req, res, next) => {
-  DOCUMENT.find(
-    {},
-    null,
-    {
-      sort: { createdAt: 1 },
-    },
-    (err, docs) => {
-      if (err) {
-        return next(err);
-      }
-      res.status(200).send(docs);
-    }
-  );
-});
-
-router.route("/:id").get((req, res, next) => {
-  DOCUMENT.findById(req.params.id, (err, go) => {
-    if (err) {
-      return next(err);
-    }
-    res.json(go);
-  });
-});
-
-//POST TEST WORKS WITH IN POSTMAN http://localhost:5000/api/document/upload
-router.post("/upload", upload.single("file"), function (req, res) {
-  const file = req.file;
-  const s3FileURL = process.env.AWS_Uploaded_File_URL_LINK;
-
-  let s3bucket = new AWS.S3({
-    accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-    secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-    region: process.env.AWS_REGION,
-  });
-
-  let params = {
-    Bucket: process.env.AWS_BUCKET_NAME,
-    Key: file.originalname,
-    Body: file.buffer,
-    ContentType: file.mimetype,
-    ACL: "public-read",
-  };
-
-  s3bucket.upload(params, function (err, data) {
-    if (err) {
-      res.status(500).json({ error: true, Message: err });
-    } else {
-      res.send({ data });
-      let newFileUploaded = {
-        description: req.body.description,
-        fileLink: s3FileURL + file.originalname,
-        s3_key: params.Key,
-      };
-      let document = new DOCUMENT(newFileUploaded);
-      document.save(function (error, newFile) {
-        if (error) {
-          throw error;
-        }
-      });
-    }
-  });
-});
-
-
-router.route("/:id").delete((req, res, next) => {
-  DOCUMENT.findByIdAndRemove(req.params.id, (err, result) => {
-    if (err) {
-      return next(err);
-    }
-
-    let s3bucket = new AWS.S3({
-      accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-      secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-      region: process.env.AWS_REGION,
-    });
-
-    let params = {
-      Bucket: process.env.AWS_BUCKET_NAME,
-      Key: result.s3_key,
-    };
-
-    s3bucket.deleteObject(params, (err, data) => {
-      if (err) {
-        console.log(err);
-      } else {
-        res.send({
-          status: "200",
-          responseType: "string",
-          response: "success",
-        });
-      }
-    });
-  });
-});
-
-module.exports = router;
